+++ tests/conftest.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-@pytest.fixture
-def anyio_backend():
-    return "asyncio"